Deprecate old upload tasks
