--- oeem_etl/tasks/__init__.py
@@ -1,1 +1,0 @@
-from .upload import UploadDatasets
